+++ arkade/src/commonMain/kotlin/com/arkade/core/fees/OnChainInput.kt
@@ -7,6 +7,6 @@ data class OnChainInput(
 ) {
     fun toCelArgs(): Map<String, Any> =
         mapOf(
-            "amount" to coin.amount,
+            "amount" to coin.amount.doubleValue(),
         )
 }
